Extract fdhwlib directly to kitcube-tools
The new automake 1.11 uses relative paths in "make dist", so copying files
from the linked directory ./src/akutil->.../fdhwlib/src/akutil fails.
To work around this issue the fdhwlib package has to be extracted to the
kitcube-tools directory.
All Makefile.am's, configure.in, README and .gitignore are adopted to this
new setup.